change the API of SetSampling Rate
the freq is in double now

+++ src/main.h
@@ -95,7 +95,7 @@ void DLL_EXPORT
 Digital_WritePeriod( int32_t *nByteArrayIdx, uint8_t *pByteArray, uint32_t nPeriod);
 
 void DLL_EXPORT
-Configure_SetSamplingRate(int32_t *nByteArrayIdx, uint8_t *pByteArray, uint32_t nfreq);
+Configure_SetSamplingRate(int32_t *nByteArrayIdx, uint8_t *pByteArray, double nfreq);
 
 /*
 void DLL_EXPORT
